all normed

+++ includes/keycode_linux.h
@@ -1,56 +1,69 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   keycode_linux.h                                    :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: calamber <calamber@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2019/12/13 19:58:42 by calamber          #+#    #+#             */
+/*   Updated: 2019/12/13 19:59:07 by calamber         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #ifndef KEYCODE_LINUX_H
 # define KEYCODE_LINUX_H
-# define FT_EOT         4
-# define KEY_ALT_LEFT	1130044187
-# define KEY_ALT_RIGHT	1146821403
-# define KEY_ALT_UP		1096489755
-# define KEY_ALT_DOWN	1113266971
-# define KEY_HOME	    4741915
-# define KEY_END        4610843
-# define KEY_BACKSPACE  (unsigned long)(127)
-# define KEY_ESCAPE			(unsigned long)(27)
-# define KEY_UP			(unsigned long)(4283163)
-# define KEY_DOWN			(unsigned long)(4348699)
-# define KEY_LEFT			(unsigned long)(4479771)
-# define KEY_RIGHT			(unsigned long)(4414235)
-# define KEY_TAB			(unsigned long)(9)
-# define KEY_Q			(unsigned long)(113)
-# define KEY_W			(unsigned long)(119)
-# define KEY_E			(unsigned long)(101)
-# define KEY_R			(unsigned long)(114)
-# define KEY_T			(unsigned long)(116)
-# define KEY_Y			(unsigned long)(121)
-# define KEY_U			(unsigned long)(117)
-# define KEY_I			(unsigned long)(105)
-# define KEY_O			(unsigned long)(111)
-# define KEY_P			(unsigned long)(112 )
-# define KEY_OPEN_BRACE			(unsigned long)(91)
-# define KEY_CLOSE_BRACE			(unsigned long)(93)
-# define KEY_BACKSLASH			(unsigned long)(92)
-# define KEY_CAPSLOCK			(unsigned long)(65509)
-# define KEY_A			(unsigned long)(97 )
-# define KEY_S			(unsigned long)(115)
-# define KEY_D			(unsigned long)(100)
-# define KEY_F			(unsigned long)(102)
-# define KEY_G			(unsigned long)(103)
-# define KEY_H			(unsigned long)(104)
-# define KEY_J			(unsigned long)(106)
-# define KEY_K			(unsigned long)(107)
-# define KEY_L			(unsigned long)(108)
-# define KEY_SEMI_COLON			(unsigned long)(59)
-# define KEY_APOSTROPHE			(unsigned long)(39)
-# define KEY_ENTER			(unsigned long)(10)
-# define KEY_Z			(unsigned long)(122)
-# define KEY_X			(unsigned long)(120)
-# define KEY_C			(unsigned long)(99)
-# define KEY_V			(unsigned long)(118)
-# define KEY_B			(unsigned long)(98)
-# define KEY_N			(unsigned long)(110)
-# define KEY_M			(unsigned long)(109)
-# define KEY_COMMA			(unsigned long)(44)
-# define KEY_PERIOD			(unsigned long)(46)
-# define KEY_SLASH			(unsigned long)(47)
-# define KEY_CTRL_LEFT			(unsigned long)(65507)
-# define KEY_SPACE			(unsigned long)(32)
-# define KEY_CTRL_RIGHT			(unsigned long)(65508)
+
+# define FT_EOT             4
+# define KEY_ALT_LEFT	    1130044187
+# define KEY_ALT_RIGHT	    1146821403
+# define KEY_ALT_UP		    1096489755
+# define KEY_ALT_DOWN	    1113266971
+# define KEY_HOME	        4741915
+# define KEY_END            4610843
+# define KEY_BACKSPACE      127
+# define KEY_ESCAPE		    27
+# define KEY_UP		        4283163
+# define KEY_DOWN		    4348699
+# define KEY_LEFT		    4479771
+# define KEY_RIGHT		    4414235
+# define KEY_TAB		    9
+# define KEY_Q		        113
+# define KEY_W		        119
+# define KEY_E		        101
+# define KEY_R		        114
+# define KEY_T		        116
+# define KEY_Y              121
+# define KEY_U	        	117
+# define KEY_I	    	    105
+# define KEY_O		        111
+# define KEY_P	        	112
+# define KEY_OPEN_BRACE		91
+# define KEY_CLOSE_BRACE	93
+# define KEY_BACKSLASH		92
+# define KEY_CAPSLOCK		65509
+# define KEY_A		        97
+# define KEY_S		        115
+# define KEY_D		        100
+# define KEY_F		        102
+# define KEY_G		        103
+# define KEY_H		        104
+# define KEY_J		        106
+# define KEY_K		        107
+# define KEY_L		        108
+# define KEY_SEMI_COLON		59
+# define KEY_APOSTROPHE		39
+# define KEY_ENTER		    10
+# define KEY_Z		        122
+# define KEY_X		        120
+# define KEY_C		        99
+# define KEY_V		        118
+# define KEY_B		        98
+# define KEY_N		        110
+# define KEY_M		        109
+# define KEY_COMMA		    44
+# define KEY_PERIOD		    46
+# define KEY_SLASH		    47
+# define KEY_CTRL_LEFT		65507
+# define KEY_SPACE		    32
+# define KEY_CTRL_RIGHT		65508
 #endif
